+++ Sources/DTCollectionViewManager/DTCollectionViewManager+Deprecated.swift
@@ -27,7 +27,7 @@ import UIKit
 
 /// Deprecated methods
 public extension DTCollectionViewManager {
-    @available(*, deprecated, message: "All cell and view events are now available inside mapping closure in register(_:handler:mapping:) method, along with type already inferred, thus making this method obsolete")
+    @available(*, deprecated, message: "All cell and view events are now available inside mapping closure in register(_:mapping:handler:) method, along with type already inferred, thus making this method obsolete")
     /// Immediately runs closure to provide access to both T and T.ModelType for `klass`.
     ///
     /// - Discussion: This is particularly useful for registering events, because near 1/3 of events don't have cell or view before they are getting run, which prevents view type from being known, and required developer to remember, which model is mapped to which cell.
@@ -39,7 +39,7 @@ public extension DTCollectionViewManager {
         closure(T.self, T.ModelType.self)
     }
     
-    @available(*, deprecated, message: "Please use handler parameter in register(_:handler:mapping:) method instead.")
+    @available(*, deprecated, message: "Please use handler parameter in register(_:mapping:handler:) method instead.")
     /// Registers `closure` to be executed, when `UICollectionView` requests `cellClass` in `UICollectionViewDataSource.collectionView(_:cellForItemAt:)` method and cell is being configured.
     ///
     /// This closure will be performed *after* cell is created and `update(with:)` method is called.
@@ -48,7 +48,7 @@ public extension DTCollectionViewManager {
         collectionDataSource?.appendReaction(for: T.self, signature: .configureCell, closure: closure)
     }
     
-    @available(*, deprecated, message: "Please use handler parameter in registerHeader(_:handler:mapping:) method instead.")
+    @available(*, deprecated, message: "Please use handler parameter in registerHeader(_:mapping:handler:) method instead.")
     /// Registers `closure` to be executed, when `UICollectionView` requests `headerClass` in `UICollectionViewDataSource.collectionView(_:viewForSupplementaryElementOf:at:)` method and header is being configured.
     ///
     /// This closure will be performed *after* header is created and `update(with:)` method is called.
@@ -60,7 +60,7 @@ public extension DTCollectionViewManager {
         configureSupplementary(T.self, ofKind: UICollectionView.elementKindSectionHeader, indexPathClosure)
     }
     
-    @available(*, deprecated, message: "Please use handler parameter in registerFooter(_:handler:mapping:) method instead.")
+    @available(*, deprecated, message: "Please use handler parameter in registerFooter(_:mapping:handler:) method instead.")
     /// Registers `closure` to be executed, when `UICollectionView` requests `footerClass` in `UICollectionViewDataSource.collectionView(_:viewForSupplementaryElementOf:at:)` method and footer is being configured.
     ///
     /// This closure will be performed *after* footer is created and `update(with:)` method is called.
@@ -72,7 +72,7 @@ public extension DTCollectionViewManager {
         configureSupplementary(T.self, ofKind: UICollectionView.elementKindSectionFooter, indexPathClosure)
     }
     
-    @available(*, deprecated, message: "Please use handler parameter in registerSupplementary(_:ofKind:handler:mapping:) method instead.")
+    @available(*, deprecated, message: "Please use handler parameter in registerSupplementary(_:ofKind:mapping:handler:) method instead.")
     /// Registers `closure` to be executed, when `UICollectionView` requests `supplementaryClass` of `kind` in `UICollectionViewDataSource.collectionView(_:viewForSupplementaryElementOf:at:)` method and supplementary view is being configured.
     ///
     /// This closure will be performed *after* supplementary view is created and `update(with:)` method is called.
@@ -81,7 +81,7 @@ public extension DTCollectionViewManager {
         collectionDataSource?.appendReaction(forSupplementaryKind: kind, supplementaryClass: T.self, signature: .configureSupplementary, closure: closure)
     }
     
-    @available(*, deprecated, message: "Please use registerSupplementary(_:kind:handler:mapping:) instead.")
+    @available(*, deprecated, message: "Please use registerSupplementary(_:kind:mapping:handler:) instead.")
     /// Registers mapping from model class to supplementary view of `supplementaryClass` type for supplementary `kind`.
     func registerNiblessSupplementary<T:ModelTransfer>(_ supplementaryClass: T.Type, forKind kind: String, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView {
         viewFactory.registerSupplementaryClass(T.self, ofKind: kind, handler: { _, _, _ in }, mapping: { mapping in
@@ -89,24 +89,24 @@ public extension DTCollectionViewManager {
             mappingBlock?(mapping)
         })
     }
-    @available(*, deprecated, message: "Please use registerHeader(_:handler:mapping:) instead.")
+    @available(*, deprecated, message: "Please use registerHeader(_:mapping:handler:) instead.")
     /// Registers mapping from model class to header view of `headerClass` type for `UICollectionElementKindSectionHeader`.
     func registerNiblessHeader<T:ModelTransfer>(_ headerClass: T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView {
-        registerHeader(T.self, handler: { _, _, _ in }, mapping: { mapping in
+        registerHeader(T.self, mapping: { mapping in
             mapping.xibName = nil
             mappingBlock?(mapping)
-        })
+        }, handler: { _, _, _ in })
     }
     
-    @available(*, deprecated, message: "Please use registerFooter(_:handler:mapping:) instead")
+    @available(*, deprecated, message: "Please use registerFooter(_:mapping:handler:) instead")
     /// Registers mapping from model class to footer view of `footerClass` type for `UICollectionElementKindSectionFooter`.
     func registerNiblessFooter<T:ModelTransfer>(_ footerClass: T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView {
-        registerFooter(T.self, handler: { _, _, _ in }, mapping: { mapping in
+        registerFooter(T.self, mapping: { mapping in
             mapping.xibName = nil
             mappingBlock?(mapping)
-        })
+        }, handler: { _, _, _ in })
     }
-    @available(*, deprecated, message: "Please use register(_:handler:mapping:) instead.")
+    @available(*, deprecated, message: "Please use register(_:mapping:handler:) instead.")
     func registerNibless<T:ModelTransfer>(_ cellClass:T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UICollectionViewCell
     {
         viewFactory.registerCellClass(T.self, handler: { _, _, _ in }, mapping: { mappingInstance in
@@ -115,45 +115,42 @@ public extension DTCollectionViewManager {
         })
     }
     
-    @available(*, deprecated, message: "Please use registerSupplementary(_:kind:handler:mapping:) and set xibName in mapping closure instead.")
+    @available(*, deprecated, message: "Please use registerSupplementary(_:kind:mapping:handler:) and set xibName in mapping closure instead.")
     /// Registers mapping from model class to supplementary view of `supplementaryClass` type with `nibName` for supplementary `kind`.
     func registerNibNamed<T:ModelTransfer>(_ nibName: String, forSupplementary supplementaryClass: T.Type, ofKind kind: String, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView
     {
         registerSupplementary(T.self, ofKind: kind,
-                              handler: { _, _, _ in },
                               mapping: { mappingInstance in
-            mappingInstance.xibName = nibName
-            mappingBlock?(mappingInstance)
-        })
+                                mappingInstance.xibName = nibName
+                                mappingBlock?(mappingInstance)
+                              }, handler: { _, _, _ in })
     }
     
-    @available(*, deprecated, message: "Please use registerHeader(_:handler:mapping:) and set xibName in mapping closure instead.")
+    @available(*, deprecated, message: "Please use registerHeader(_:mapping:handler:) and set xibName in mapping closure instead.")
     /// Registers mapping from model class to supplementary view of `headerClass` type with `nibName` for UICollectionElementKindSectionHeader.
     func registerNibNamed<T:ModelTransfer>(_ nibName: String, forHeader headerClass: T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView
     {
         registerSupplementary(T.self,
                               ofKind: UICollectionView.elementKindSectionHeader,
-                              handler: { _, _, _ in },
                               mapping: { mappingInstance in
-            mappingInstance.xibName = nibName
-            mappingBlock?(mappingInstance)
-        })
+                                mappingInstance.xibName = nibName
+                                mappingBlock?(mappingInstance)
+                              }, handler: { _, _, _ in })
     }
     
-    @available(*, deprecated, message: "Please use registerFooter(_:handler:mapping:) and set xibName in mapping closure instead.")
+    @available(*, deprecated, message: "Please use registerFooter(_:mapping:handler:) and set xibName in mapping closure instead.")
     /// Registers mapping from model class to supplementary view of `footerClass` type with `nibName` for UICollectionElementKindSectionFooter.
     func registerNibNamed<T:ModelTransfer>(_ nibName: String, forFooter footerClass: T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView
     {
         registerSupplementary(T.self,
                               ofKind: UICollectionView.elementKindSectionFooter,
-                              handler: { _, _, _ in },
                               mapping: { mappingInstance in
-            mappingInstance.xibName = nibName
-            mappingBlock?(mappingInstance)
-        })
+                                mappingInstance.xibName = nibName
+                                mappingBlock?(mappingInstance)
+                              }, handler: { _, _, _ in })
     }
     
-    @available(*, deprecated, message: "Please use register(_:handler:mapping:) and set xibName in mapping closure instead.")
+    @available(*, deprecated, message: "Please use register(_:mapping:handler:) and set xibName in mapping closure instead.")
     /// Registers nib with `nibName` mapping from model class to `cellClass`.
     func registerNibNamed<T:ModelTransfer>(_ nibName: String, for cellClass: T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UICollectionViewCell
     {