+++ Sources/DTCollectionViewManager/DTCollectionViewManager+Registration.swift
@@ -28,26 +28,26 @@ import DTModelStorage
 
 /// Upgrade shims for easier API upgrading
 public extension DTCollectionViewManager {
-    @available(*, unavailable, renamed: "register(_:handler:mapping:)")
-    /// This method is unavailable, please use `register(_:handler:mapping:)` as a replacement.
+    @available(*, unavailable, renamed: "register(_:mapping:handler:)")
+    /// This method is unavailable, please use `register(_:mapping:handler:)` as a replacement.
     func register<T:ModelTransfer>(_ cellClass:T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UICollectionViewCell
     {
     }
     
-    @available(*, unavailable, renamed: "registerSupplementary(_:ofKind:handler:mapping:)")
-    /// This method is unavailable, please use `registerSupplementary(_:handler:mapping:)` as a replacement.
+    @available(*, unavailable, renamed: "registerSupplementary(_:ofKind:mapping:handler:)")
+    /// This method is unavailable, please use `registerSupplementary(_:mapping:handler:)` as a replacement.
     func registerSupplementary<T:ModelTransfer>(_ supplementaryClass: T.Type, forKind kind: String, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView
     {
     }
     
-    @available(*, unavailable, renamed: "registerHeader(_:handler:mapping:)")
-    /// This method is unavailable, please use `registerHeader(_:handler:mapping:)` as a replacement.
+    @available(*, unavailable, renamed: "registerHeader(_:mapping:handler:)")
+    /// This method is unavailable, please use `registerHeader(_:mapping:handler:)` as a replacement.
     func registerHeader<T:ModelTransfer>(_ headerClass : T.Type, mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UICollectionReusableView {
         
     }
     
-    @available(*, unavailable, renamed: "registerFooter(_:handler:mapping:)")
-    /// This method is unavailable, please use `registerFooter(_:handler:mapping:)` as a replacement.
+    @available(*, unavailable, renamed: "registerFooter(_:mapping:handler:)")
+    /// This method is unavailable, please use `registerFooter(_:mapping:handler:)` as a replacement.
     func registerFooter<T:ModelTransfer>(_ footerClass: T.Type,
                                               mappingBlock: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T:UICollectionReusableView {
         
@@ -58,10 +58,11 @@ extension DTCollectionViewManager {
     /// Registers mapping for `cellClass`. Mapping will automatically check for nib with the same name as `cellClass` and register it, if it is found. If cell is designed in storyboard, please set `mapping.cellRegisteredByStoryboard` property to `true` inside of `mapping` closure.
     /// - Parameters:
     ///   - cellClass: UICollectionViewCell subclass type, conforming to `ModelTransfer` protocol.
-    ///   - handler: configuration closure, that is run when cell is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    ///   - handler: configuration closure, that is run when cell is dequeued.
     /// - Note: `handler` closure is called before `update(with:)` method.
-    open func register<T:ModelTransfer>(_ cellClass:T.Type, handler: @escaping (T, T.ModelType, IndexPath) -> Void = { _, _, _ in }, mapping: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil) where T: UICollectionViewCell
+    open func register<T:ModelTransfer>(_ cellClass:T.Type, mapping: ((ViewModelMapping<T, T.ModelType>) -> Void)? = nil,
+                                        handler: @escaping (T, T.ModelType, IndexPath) -> Void = { _, _, _ in }) where T: UICollectionViewCell
     {
         viewFactory.registerCellClass(T.self, handler: handler, mapping: mapping)
     }
@@ -70,9 +71,9 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - cellClass: UICollectionViewCell to register
     ///   - modelType: Model type, which is mapped to `cellClass`.
-    ///   - handler: configuration closure, that is run when cell is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
-    open func register<Cell: UICollectionViewCell, Model>(_ cellClass: Cell.Type, for modelType: Model.Type, handler: @escaping (Cell, Model, IndexPath) -> Void, mapping: ((ViewModelMapping<Cell, Model>) -> Void)? = nil) {
+    ///   - handler: configuration closure, that is run when cell is dequeued.
+    open func register<Cell: UICollectionViewCell, Model>(_ cellClass: Cell.Type, for modelType: Model.Type, mapping: ((ViewModelMapping<Cell, Model>) -> Void)? = nil, handler: @escaping (Cell, Model, IndexPath) -> Void) {
         viewFactory.registerCellClass(cellClass, modelType, handler: handler, mapping: mapping)
     }
 
@@ -81,12 +82,12 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - headerClass: UICollectionReusableView class to register
     ///   - modelType: Model type, which is mapped to `headerClass`.
-    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     /// - Note: `handler` closure is called before `update(with:)` method.
     open func registerHeader<View:ModelTransfer>(_ headerClass : View.Type,
-                                              handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in },
-                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil) where View: UICollectionReusableView
+                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil,
+                                              handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in }) where View: UICollectionReusableView
     {
         viewFactory.registerSupplementaryClass(View.self,
                                                ofKind: UICollectionView.elementKindSectionHeader,
@@ -99,10 +100,13 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - headerClass: UICollectionReusableView class to register
     ///   - modelType: Model type, which is mapped to `headerClass`.
-    ///   - handler: configuration closure, that is run when header is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
-    open func registerHeader<View:UICollectionReusableView, Model>(_ headerClass: View.Type, for modelType: Model.Type, handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in }, mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil) {
-        registerSupplementary(View.self, for: Model.self, ofKind: UICollectionView.elementKindSectionHeader, handler: handler, mapping: mapping)
+    ///   - handler: configuration closure, that is run when header is dequeued.
+    open func registerHeader<View:UICollectionReusableView, Model>(_ headerClass: View.Type,
+                                                                   for modelType: Model.Type,
+                                                                   mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil,
+                                                                   handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in }) {
+        registerSupplementary(View.self, for: Model.self, ofKind: UICollectionView.elementKindSectionHeader, mapping: mapping, handler: handler)
     }
     
     /// Registers mapping for `footerClass`. `UICollectionView.elementKindSectionFooter` is used as a supplementary kind. Mapping will automatically check for nib with the same name as `footerClass` and register it, if it is found.
@@ -110,12 +114,12 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - footerClass: UICollectionReusableView class to register
     ///   - modelType: Model type, which is mapped to `footerClass`.
-    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     /// - Note: `handler` closure is called before `update(with:)` method.
     open func registerFooter<View:ModelTransfer>(_ footerClass: View.Type,
-                                              handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in },
-                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil) where View:UICollectionReusableView
+                                              mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil,
+                                              handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in }) where View:UICollectionReusableView
     {
         viewFactory.registerSupplementaryClass(View.self,
                                                ofKind: UICollectionView.elementKindSectionFooter,
@@ -128,20 +132,28 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - footerClass: UICollectionReusableView class to register
     ///   - modelType: Model type, which is mapped to `footerClass`.
-    ///   - handler: configuration closure, that is run when footer is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
-    open func registerFooter<View:UICollectionReusableView, Model>(_ footerClass: View.Type, for modelType: Model.Type, handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in }, mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil) {
-        registerSupplementary(View.self, for: Model.self, ofKind: UICollectionView.elementKindSectionFooter, handler: handler, mapping: mapping)
+    ///   - handler: configuration closure, that is run when footer is dequeued.
+    open func registerFooter<View:UICollectionReusableView, Model>(_ footerClass: View.Type,
+                                                                   for modelType: Model.Type,
+                                                                   mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil,
+                                                                   handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in })
+    {
+        registerSupplementary(View.self, for: Model.self, ofKind: UICollectionView.elementKindSectionFooter, mapping: mapping, handler: handler)
     }
     
-    /// Registers mapping from model class to suppplementary view of `supplementaryClass` type for supplementary `kind`.
-    ///
-    /// Method will automatically check for nib with the same name as `supplementaryClass`. If it exists - nib will be registered instead of class.
+    /// Registers mapping for `footerClass`.  Mapping will automatically check for nib with the same name as `supplementaryClass` and register it, if it is found.
+    /// If supplementary view is designed in storyboard, please set `mapping.supplementaryRegisteredByStoryboard` property to `true` inside of `mapping` closure.
+    /// - Parameters:
+    ///   - supplementaryClass: UICollectionReusableView class to register
+    ///   - kind: supplementary view kind
+    ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
+    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     /// - Note: `handler` closure is called before `update(with:)` method.
     open func registerSupplementary<View:ModelTransfer>(_ supplementaryClass: View.Type,
                                                      ofKind kind: String,
-                                                     handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in },
-                                                     mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil) where View:UICollectionReusableView
+                                                     mapping: ((ViewModelMapping<View, View.ModelType>) -> Void)? = nil,
+                                                     handler: @escaping (View, View.ModelType, IndexPath) -> Void = { _, _, _ in }) where View:UICollectionReusableView
     {
         viewFactory.registerSupplementaryClass(View.self, ofKind: kind, handler: handler, mapping: mapping)
     }
@@ -151,9 +163,14 @@ extension DTCollectionViewManager {
     /// - Parameters:
     ///   - footerClass: UICollectionReusableView class to register
     ///   - modelType: Model type, which is mapped to `supplementaryClass`.
-    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
     ///   - mapping: mapping configuration closure, executed before any registration or dequeue is performed.
-    open func registerSupplementary<View:UICollectionReusableView, Model>(_ supplementaryClass: View.Type, for modelType: Model.Type, ofKind kind: String, handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in }, mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil) {
+    ///   - handler: configuration closure, that is run when supplementary view is dequeued.
+    open func registerSupplementary<View:UICollectionReusableView, Model>(_ supplementaryClass: View.Type,
+                                                                          for modelType: Model.Type,
+                                                                          ofKind kind: String,
+                                                                          mapping: ((ViewModelMapping<View, Model>) -> Void)? = nil,
+                                                                          handler: @escaping (View, Model, IndexPath) -> Void = { _, _, _ in })
+    {
         viewFactory.registerSupplementaryClass(supplementaryClass, modelType, ofKind: kind, handler: handler, mapping: mapping)
     }